Proyecto/backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
drf-spectacular==0.27.0
mysqlclient==2.2.0
python-dotenv==1.0.0
Pillow==10.0.0
pymysql
cryptography

